[♻️refactor]: interface 파일 이동 리팩토링
